moved domain tests to their own subfolder
